fixed errors within calib pixel, find clus, cluster charge cut,validated results !!

+++ src/serial/CUDADataFormats/SiPixelDigisSoA.jl
@@ -2,11 +2,11 @@ module CUDADataFormatsSiPixelDigiInterfaceSiPixelDigisSoA
   # Structure to hold a constant view of device data
   potato = 123
   struct DeviceConstView
-    xx::Vector{UInt16}         # X-coordinates of pixels
-    yy::Vector{UInt16}         # Y-coordinates of pixels
-    adc::Vector{UInt16}        # ADC values of pixels
-    module_ind::Vector{UInt16}  # Module indices of pixels
-    clus::Vector{UInt32}       # Cluster indices of pixels
+    xx::Vector{Int16}         # X-coordinates of pixels
+    yy::Vector{Int16}         # Y-coordinates of pixels
+    adc::Vector{Int16}        # ADC values of pixels
+    module_ind::Vector{Int16}  # Module indices of pixels
+    clus::Vector{Int32}       # Cluster indices of pixels
   end
 
   # Structure to hold SiPixel digis data
@@ -14,10 +14,10 @@ module CUDADataFormatsSiPixelDigiInterfaceSiPixelDigisSoA
   mutable struct SiPixelDigisSoA
       pdigi_d::Vector{UInt32}      # Digis data
       raw_id_arr_d::Vector{UInt32}   # Raw ID array
-      xx_d::Vector{UInt16}         # Local X-coordinates of each pixel
-      yy_d::Vector{UInt16}         # Local Y-coordinates of each pixel
-      adc_d::Vector{UInt16}        # ADC values of each pixel
-      module_ind_d::Vector{UInt16}  # Module IDs of each pixel
+      xx_d::Vector{Int16}         # Local X-coordinates of each pixel
+      yy_d::Vector{Int16}         # Local Y-coordinates of each pixel
+      adc_d::Vector{Int32}        # ADC values of each pixel
+      module_ind_d::Vector{Int16}  # Module IDs of each pixel
       clus_d::Vector{Int32}       # Cluster IDs of each pixel
       view_d::DeviceConstView      # "Me" pointer, a constant view of the device data
       n_modules_h::UInt32           # Number of modules
@@ -32,11 +32,11 @@ module CUDADataFormatsSiPixelDigiInterfaceSiPixelDigisSoA
       """
       function SiPixelDigisSoA(maxFedWords::Int)
           # Uninitialized arrays of the specified size
-          xx_d = Vector{UInt16}(undef, maxFedWords)
-          yy_d = Vector{UInt16}(undef, maxFedWords)
-          adc_d = Vector{UInt16}(undef, maxFedWords)
-          module_ind_d = Vector{UInt16}(undef, maxFedWords)
-          clus_d = Vector{UInt32}(undef, maxFedWords)
+          xx_d = Vector{Int16}(undef, maxFedWords)
+          yy_d = Vector{Int16}(undef, maxFedWords)
+          adc_d = Vector{Int32}(undef, maxFedWords)
+          module_ind_d = Vector{Int16}(undef, maxFedWords)
+          clus_d = Vector{Int32}(undef, maxFedWords)
           pdigi_d = Vector{UInt32}(undef, maxFedWords)
           raw_id_arr_d = Vector{UInt32}(undef, maxFedWords)
